DAOS-3020: test: Use clush instead of pdsh in ftest.sh
Also make ftest.sh work better outside of Jenkins.

+++ ftest.sh
@@ -70,8 +70,8 @@ cleanup() {
     restore_dist_files "${yaml_files[@]}"
     i=5
     while [ $i -gt 0 ]; do
-        pdsh -l "${REMOTE_ACCT:-jenkins}" -R ssh -S \
-             -w "$(IFS=','; echo "${nodes[*]}")" "set -x
+        if ! clush "${CLUSH_ARGS[@]}" -B -l "${REMOTE_ACCT:-jenkins}" -R ssh \
+             -S -w "$(IFS=','; echo "${nodes[*]}")" "set -x
         if grep /mnt/daos /proc/mounts; then
             if ! sudo umount /mnt/daos; then
                 echo \"During shutdown, failed to unmount /mnt/daos.  \"\
@@ -80,26 +80,27 @@ cleanup() {
         fi
         x=0
         sudo sed -i -e \"/added by ftest.sh/d\" /etc/fstab
-        while [ \$x -lt 30 ] &&
-              grep $DAOS_BASE /proc/mounts &&
-              ! sudo umount $DAOS_BASE; do
-            sleep 1
-            let x+=1
-        done
-        if grep $DAOS_BASE /proc/mounts; then
-            echo \"Failed to unmount $DAOS_BASE\"
-            exit 1
-        fi
-        if [ -d $DAOS_BASE ] && ! sudo rmdir $DAOS_BASE; then
-            echo \"Failed to remove $DAOS_BASE\"
-            if [ -d $DAOS_BASE ]; then
-                ls -l $DAOS_BASE
-            else
-                echo \"because it doesnt exist\"
+        if [ -n \"$DAOS_BASE\" ]; then
+            while [ \$x -lt 30 ] &&
+                  grep $DAOS_BASE /proc/mounts &&
+                  ! sudo umount $DAOS_BASE; do
+                sleep 1
+                let x+=1
+            done
+            if grep $DAOS_BASE /proc/mounts; then
+                echo \"Failed to unmount $DAOS_BASE\"
+                exit 1
             fi
-            exit 1
-        fi" 2>&1 | dshbak -c
-        if [ "${PIPESTATUS[0]}" = 0 ]; then
+            if [ -d $DAOS_BASE ] && ! sudo rmdir $DAOS_BASE; then
+                echo \"Failed to remove $DAOS_BASE\"
+                if [ -d $DAOS_BASE ]; then
+                    ls -l $DAOS_BASE
+                else
+                    echo \"because it doesnt exist\"
+                fi
+                exit 1
+            fi
+        fi"; then
             i=0
         fi
         ((i-=1))
@@ -136,8 +137,10 @@ mkdir -p src/tests/ftest/avocado/job-results
 
 trap 'set +e; cleanup' EXIT
 
+CLUSH_ARGS=($CLUSH_ARGS)
+
 DAOS_BASE=${SL_PREFIX%/install}
-if ! pdsh -l "${REMOTE_ACCT:-jenkins}" -R ssh -S \
+if ! clush "${CLUSH_ARGS[@]}" -B -l "${REMOTE_ACCT:-jenkins}" -R ssh -S \
     -w "$(IFS=','; echo "${nodes[*]}")" "set -ex
 ulimit -c unlimited
 if [ \"\${HOSTNAME%%%%.*}\" != \"${nodes[0]}\" ]; then
@@ -183,7 +186,7 @@ rm -rf \"${TEST_TAG_DIR:?}/\"
 mkdir -p \"$TEST_TAG_DIR/\"
 if [ -z \"\$JENKINS_URL\" ]; then
     exit 0
-fi" 2>&1 | dshbak -c; then
+fi"; then
     echo "Cluster setup (i.e. provisioning) failed"
     exit 1
 fi
@@ -194,7 +197,7 @@ args+=" $*"
 
 # shellcheck disable=SC2029
 # shellcheck disable=SC2086
-if ! ssh $SSH_KEY_ARGS "${REMOTE_ACCT:-jenkins}"@"${nodes[0]}" "set -ex
+if ! ssh $SSH_KEY_ARGS ${REMOTE_ACCT:-jenkins}@"${nodes[0]}" "set -ex
 ulimit -c unlimited
 rm -rf $DAOS_BASE/install/tmp
 mkdir -p $DAOS_BASE/install/tmp
